requirements.txt
PyYAML==5.1.2
tabulate==0.8.6
ruamel.yaml==0.16.10
